Fix broken result check
In the case of a list, Anders made it so that the list would be checked for all 0's (success) and convert it from a list into a scalar 0.  However, he forgot to change the next elif to an if, so the result-checking gets skipped over after an all-successful result list gets changed to a scalar 0.  This fix should take care of that, but is untested.