Changes that I hope will make the selection of field maps more reliable.
The indenting in the previous version was strange, and I think
testing for success may have been wrong in some cases, in that
warnings about failure could be issued before the search was
complete.

dcm2bids doesn't create lists for the IntendedFor field when
a single target is provided. This breaks the original tests,
which used a ' '.join syntax and thus created strings
separated by spaces when provided with a string. This
meant the correct field map was never found.

I have forced the result into a list if it isn't already,
fixed the indenting and corrected the test for success.

The join option could be used because it allows paths without
a session field to be matched as substrings.

Recent versions of dcm2bids include the session field.